move old comp to legacy dir
