+++ packages/keychain/src/components/Auth/BannerImage.tsx
