+++ packages/keychain/src/components/Auth/PortalFooter/SessionDetails.tsx
