+++ packages/keychain/src/components/Auth/RegistrationLink.tsx
