+++ packages/keychain/src/components/Auth/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Authenticate";
+export * from "./Login";
+export * from "./Signup";
