+++ packages/keychain/src/components/legacy/signup/BannerImage.tsx
