+++ packages/keychain/src/components/legacy/signup/Continue.tsx
@@ -1,6 +1,6 @@
 import { Text, VStack, Link, StyleProps } from "@chakra-ui/react";
 import { Header } from "components/Header";
-import Container from "components/Container";
+import Container from "components/legacy/Container";
 import { SparklesDuoIcon } from "@cartridge/ui";
 
 const Continue = ({ ...rest }: StyleProps) => {
