+++ packages/keychain/src/components/legacy/signup/L1Connect.tsx
@@ -1,7 +1,6 @@
 import { useState, useCallback } from "react";
 import {
   Flex,
-  Box,
   Text,
   Link,
   Circle,
