+++ packages/keychain/src/components/legacy/signup/index.tsx
@@ -1,9 +1,7 @@
 export * from "./Form";
-export * from "./Signup";
 export * from "./Quests";
 export * from "./Pending";
 export * from "./L1Connect";
 export * from "./L2Connect";
 export * from "./StarterPack";
 export * from "./OutOfStock";
-export * from "./Authenticate";
