--- packages/keychain/src/components/v2/Auth/index.tsx
@@ -1,2 +1,0 @@
-export * from "./Signup";
-export * from "./Login";
