--- packages/keychain/src/components/v2/index.ts
@@ -1,1 +1,0 @@
-export * from "./Auth";
