+++ packages/keychain/src/pages/claim/[gameId].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import type { NextPage } from "next";
-import { StarterPack } from "components/signup";
+import { StarterPack } from "components/legacy/signup";
 import { useMemo } from "react";
 import Controller from "utils/controller";
