+++ packages/keychain/src/pages/signup/index.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import type { NextPage } from "next";
-import { Signup as SignupComponent } from "components/v2";
+import { Signup as SignupComponent } from "components";
 import { useMemo } from "react";
 import Controller from "utils/controller";
